cli/pyproject.toml
[tool.poetry]
name = "trainloop-cli"
version = "0.3.16"
description = "TrainLoop command-line interface"
authors = ["Mason Pierce <masonpierce@trainloop.ai>"]
readme = "README.md"

# Include trainloop_cli package and UI files
packages = [{ include = "trainloop_cli" }]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
black = "^23.3.0"
flake8 = "^6.0"


[tool.poetry.scripts]
# exposes a `trainloop` CLI entrypoint:
trainloop = "trainloop_cli.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
